The Background of this level is now an Entity
To change the background, we just have to change the ID: inside the JSON
String. This commit is currently have it load level 2, as not to confuse
anyone with the previous stack background, which is still behind it on
the HTML page (harccoded still). We create Long levelels horizontally by
using the Length setting for each level inside the JSON STRING/FILE for
each level. There are now 10 levels possible we can use