Implemented coins from boxes to increment points (Has bugs)
I implemented when a coin is popped from a box, it can have the value
"POPPED" passed in as the last value during creation of the entity. This
will determine if it was popped from a box, and will animate and should
accumulate the points automaticly. ITS SUPER BUGGY STILL